+++ tests/test_mqtt.py
@@ -32,7 +32,11 @@ import switchbot_mqtt
 @pytest.mark.parametrize("mqtt_host", ["mqtt-broker.local"])
 @pytest.mark.parametrize("mqtt_port", [1833])
 @pytest.mark.parametrize("retry_count", [3, 21])
-def test__run(caplog, mqtt_host, mqtt_port, retry_count):
+@pytest.mark.parametrize(
+    "device_passwords",
+    [{}, {"11:22:33:44:55:66": "password", "aa:bb:cc:dd:ee:ff": "secret"}],
+)
+def test__run(caplog, mqtt_host, mqtt_port, retry_count, device_passwords):
     with unittest.mock.patch(
         "paho.mqtt.client.Client"
     ) as mqtt_client_mock, caplog.at_level(logging.DEBUG):
@@ -42,10 +46,12 @@ def test__run(caplog, mqtt_host, mqtt_port, retry_count):
             mqtt_username=None,
             mqtt_password=None,
             retry_count=retry_count,
+            device_passwords=device_passwords,
         )
     mqtt_client_mock.assert_called_once_with(
         userdata=switchbot_mqtt._MQTTCallbackUserdata(
-            retry_count=retry_count, device_passwords={}
+            retry_count=retry_count,
+            device_passwords=device_passwords,
         )
     )
     assert not mqtt_client_mock().username_pw_set.called
@@ -104,6 +110,7 @@ def test__run_authentication(mqtt_host, mqtt_port, mqtt_username, mqtt_password)
             mqtt_username=mqtt_username,
             mqtt_password=mqtt_password,
             retry_count=7,
+            device_passwords={},
         )
     mqtt_client_mock.assert_called_once_with(
         userdata=switchbot_mqtt._MQTTCallbackUserdata(
@@ -127,6 +134,7 @@ def test__run_authentication_missing_username(mqtt_host, mqtt_port, mqtt_passwor
                 mqtt_username=None,
                 mqtt_password=mqtt_password,
                 retry_count=3,
+                device_passwords={},
             )